+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: smelicha <smelicha@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/10 23:49:36 by smelicha          #+#    #+#             */
-/*   Updated: 2023/05/11 20:04:19 by smelicha         ###   ########.fr       */
+/*   Updated: 2023/05/14 20:30:34 by smelicha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,8 +52,8 @@ int		ft_print_pointer(t_data *data);
 int		ft_print_string(t_data *data);
 int		ft_print_unsigned_int(t_data *data);
 int		ft_char_comp(char c, char *set);
-char	*ft_utoa(size_t n);
-char	*ft_hextoa(size_t n, t_data *data);
+char	*ft_utoa(long unsigned n);
+char	*ft_hextoa(long unsigned n, t_data *data);
 void	ft_print_pad(t_data *data);
 
 #endif
